refactor: useCardCompany를 쪼개어 useCardCompanyState, useCardCompanyInputHandler 두 개의 커스텀 훅으로 분리
