+++ src/App.tsx
@@ -4,13 +4,13 @@ import CardNumber from "./CardNumber/CardNumber";
 import CardExpirationDate from "./CardExpirationDate/CardExpirationDate";
 import CardCvcNumber from "./CardCvcNumber/CardCvcNumber";
 import PreviewCardLayout from "./components/PreviewCard/PreviewCardLayout";
-import useCvcNumberState from "./hooks/useCvcNumber/useCvcNumberState";
-import useCardNumbersState from "./hooks/useCardNumber/useCardNumberState";
 import CardCompanyPicker from "./CardCompany/CardCompany";
-import useCardCompany from "./hooks/useCardCompany";
 import useCardPassword from "./hooks/useCardPassword";
 import CardPassword from "./CardPassword/CardPassword";
+import useCardNumbersState from "./hooks/useCardNumber/useCardNumberState";
 import useExpirationDateState from "./hooks/useExpirationDate/useExpirationDateState";
+import useCvcNumberState from "./hooks/useCvcNumber/useCvcNumberState";
+import useCardCompanyState from "./hooks/useCardCompany/useCardCompanyState";
 
 function App() {
   const { cardNumbers, cardType, cardNumbersError, handleChange } =
@@ -19,7 +19,7 @@ function App() {
     useExpirationDateState();
   const { cvcNumbers, cvcNumbersError, cvcNumbersValidate } =
     useCvcNumberState();
-  const { selectedCompany, selectCompany } = useCardCompany();
+  const { selectedCompany, selectCompany } = useCardCompanyState();
   const { password, passwordError, passwordValidate } = useCardPassword();
   const [step, setStep] = useState(1);
